requirements.txt
pydantic>=2.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-ollama>=1.0.0
langchain-core>=0.1.0
requests>=2.31.0
fastapi>=0.110.0
uvicorn>=0.29.0
